backend/requirements.txt
fastapi
uvicorn[standard]
opencv-python
numpy
pytesseract
Pillow
python-multipart
pandas
rapidocr-onnxruntime
onnxruntime
paddleocr
